Remove `canvas` dependencies to allow use with gatsby
Although it is possible to use `trianglify` in `react`,
it is not possible to use it with a static site generator
like `gatsby` because `canvas` is delivered as a binary.
This commit removes all `canvas`-related functionality.
Trianglify now returns a list of polygons and options.
Using those polygons and options, it is possible to build
JSX svg components that can be used on a static website.